Move client code to client folder
Even  if it is used in the server.

We should be able to split the client folder and use it independently of the
server later.
