+++ lib/client/components/app/index.js
