+++ lib/server/routes.js
@@ -4,9 +4,9 @@ import React from 'react'
 import {renderToString} from 'react-dom/server'
 import {match, RoutingContext} from 'react-router'
 
-import routes from '../routes'
+import routes from '../client/routes'
 import views from './views'
-import api from '../api'
+import api from '../client/api'
 
 const router = express.Router()
 export default router
